chore(docs): Fix the docs as per feedback
