+++ .github/workflows/ci.yml
@@ -1,0 +1,76 @@
+name: Build
+
+on:
+  push:
+    branches: [main]
+  pull_request:
+    branches: [main]
+
+jobs:
+  build:
+    name: Build
+    runs-on: ubuntu-latest
+    env:
+      GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
+      NODE_AUTH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+    outputs:
+      version: ${{ steps.version.outputs.version }}
+      version-major: ${{ steps.version.outputs.version-major-only }}
+      version-minor: ${{ format('{0}.{1}', steps.version.outputs.version-major-only, steps.version.outputs.version-minor-only) }}
+      version-patch: ${{ format('{0}.{1}.{2}', steps.version.outputs.version-major-only, steps.version.outputs.version-minor-only, steps.version.outputs.version-patch-only) }}
+    permissions:
+      contents: read
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+      - name: Generate next version
+        id: version
+        uses: conventional-actions/next-version@v1
+      - name: Setup Node
+        uses: actions/setup-node@v3
+        with:
+          node-version: 18
+          cache: "npm"
+      - name: Setup npmrc for ketch-sdk
+        uses: conventional-actions/setup-npmrc@v1
+        with:
+          always-auth: true
+          registry-url: https://registry.npmjs.com/
+          scope: ketch-sdk
+          token: ${{ secrets.NPM_TOKEN }}
+      - name: Setup npmrc for ketch-com
+        uses: conventional-actions/setup-npmrc@v1
+        with:
+          always-auth: true
+          scope: ketch-com
+          token: ${{ secrets.GH_TOKEN }}
+      - name: Setup Node version
+        uses: conventional-actions/write-version@v1
+        with:
+          output_path: package.json
+          version: ${{ format('{0}.{1}.{2}', steps.version.outputs.version-major-only, steps.version.outputs.version-minor-only, steps.version.outputs.version-patch-only) }}
+      - name: Install dependencies
+        run: npm ci
+      - name: Run all targets
+        run: npm run all
+      - name: Publish
+        if: github.ref == format('refs/heads/{0}', github.event.repository.default_branch)
+        run: npm publish --access public
+
+  release:
+    name: Release
+    if: github.ref == format('refs/heads/{0}', github.event.repository.default_branch)
+    needs:
+      - build
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+    steps:
+      - name: Create Release
+        uses: conventional-actions/create-release@v1
+        with:
+          tag_name: ${{ needs.build.outputs.version }}
+          artifacts: '*'
