+++ docs/assets/search.js
@@ -1,1 +1,1 @@
-window.searchData = JSON.parse("{\"rows\":[{\"kind\":2,\"name\":\"index\",\"url\":\"modules/index.html\",\"classes\":\"\"},{\"kind\":64,\"name\":\"handleRequest\",\"url\":\"functions/index.handleRequest.html\",\"classes\":\"\",\"parent\":\"index\"},{\"kind\":2,\"name\":\"types\",\"url\":\"modules/types.html\",\"classes\":\"\"},{\"kind\":8,\"name\":\"RequestStatus\",\"url\":\"enums/types.RequestStatus.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":16,\"name\":\"UnknownRequestStatus\",\"url\":\"enums/types.RequestStatus.html#UnknownRequestStatus\",\"classes\":\"\",\"parent\":\"types.RequestStatus\"},{\"kind\":16,\"name\":\"PendingRequestStatus\",\"url\":\"enums/types.RequestStatus.html#PendingRequestStatus\",\"classes\":\"\",\"parent\":\"types.RequestStatus\"},{\"kind\":16,\"name\":\"InProgressRequestStatus\",\"url\":\"enums/types.RequestStatus.html#InProgressRequestStatus\",\"classes\":\"\",\"parent\":\"types.RequestStatus\"},{\"kind\":16,\"name\":\"CompletedRequestStatus\",\"url\":\"enums/types.RequestStatus.html#CompletedRequestStatus\",\"classes\":\"\",\"parent\":\"types.RequestStatus\"},{\"kind\":16,\"name\":\"CancelledRequestStatus\",\"url\":\"enums/types.RequestStatus.html#CancelledRequestStatus\",\"classes\":\"\",\"parent\":\"types.RequestStatus\"},{\"kind\":16,\"name\":\"DeniedRequestStatus\",\"url\":\"enums/types.RequestStatus.html#DeniedRequestStatus\",\"classes\":\"\",\"parent\":\"types.RequestStatus\"},{\"kind\":8,\"name\":\"RequestStatusReason\",\"url\":\"enums/types.RequestStatusReason.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":16,\"name\":\"UnknownRequestStatusReason\",\"url\":\"enums/types.RequestStatusReason.html#UnknownRequestStatusReason\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":16,\"name\":\"NeedUserVerificationRequestStatusReason\",\"url\":\"enums/types.RequestStatusReason.html#NeedUserVerificationRequestStatusReason\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":16,\"name\":\"Requested\",\"url\":\"enums/types.RequestStatusReason.html#Requested\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":16,\"name\":\"InsufficientIdentification\",\"url\":\"enums/types.RequestStatusReason.html#InsufficientIdentification\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":16,\"name\":\"Executed\",\"url\":\"enums/types.RequestStatusReason.html#Executed\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":16,\"name\":\"SuspectedFraudRequestStatusReason\",\"url\":\"enums/types.RequestStatusReason.html#SuspectedFraudRequestStatusReason\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":16,\"name\":\"InsufficientVerificationRequestStatusReason\",\"url\":\"enums/types.RequestStatusReason.html#InsufficientVerificationRequestStatusReason\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":16,\"name\":\"NoMatchRequestStatusReason\",\"url\":\"enums/types.RequestStatusReason.html#NoMatchRequestStatusReason\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":16,\"name\":\"ClaimNotCoveredRequestStatusReason\",\"url\":\"enums/types.RequestStatusReason.html#ClaimNotCoveredRequestStatusReason\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":16,\"name\":\"OutsideJurisdictionRequestStatusReason\",\"url\":\"enums/types.RequestStatusReason.html#OutsideJurisdictionRequestStatusReason\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":16,\"name\":\"TooManyRequestsRequestStatusReason\",\"url\":\"enums/types.RequestStatusReason.html#TooManyRequestsRequestStatusReason\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":16,\"name\":\"OtherRequestStatusReason\",\"url\":\"enums/types.RequestStatusReason.html#OtherRequestStatusReason\",\"classes\":\"\",\"parent\":\"types.RequestStatusReason\"},{\"kind\":8,\"name\":\"Kind\",\"url\":\"enums/types.Kind.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":16,\"name\":\"AccessRequest\",\"url\":\"enums/types.Kind.html#AccessRequest\",\"classes\":\"\",\"parent\":\"types.Kind\"},{\"kind\":16,\"name\":\"AccessResponse\",\"url\":\"enums/types.Kind.html#AccessResponse\",\"classes\":\"\",\"parent\":\"types.Kind\"},{\"kind\":16,\"name\":\"AccessStatusEvent\",\"url\":\"enums/types.Kind.html#AccessStatusEvent\",\"classes\":\"\",\"parent\":\"types.Kind\"},{\"kind\":16,\"name\":\"DeleteRequest\",\"url\":\"enums/types.Kind.html#DeleteRequest\",\"classes\":\"\",\"parent\":\"types.Kind\"},{\"kind\":16,\"name\":\"DeleteResponse\",\"url\":\"enums/types.Kind.html#DeleteResponse\",\"classes\":\"\",\"parent\":\"types.Kind\"},{\"kind\":16,\"name\":\"DeleteStatusEvent\",\"url\":\"enums/types.Kind.html#DeleteStatusEvent\",\"classes\":\"\",\"parent\":\"types.Kind\"},{\"kind\":256,\"name\":\"Identity\",\"url\":\"interfaces/types.Identity.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":1024,\"name\":\"identitySpace\",\"url\":\"interfaces/types.Identity.html#identitySpace\",\"classes\":\"\",\"parent\":\"types.Identity\"},{\"kind\":1024,\"name\":\"identityFormat\",\"url\":\"interfaces/types.Identity.html#identityFormat\",\"classes\":\"\",\"parent\":\"types.Identity\"},{\"kind\":1024,\"name\":\"identityValue\",\"url\":\"interfaces/types.Identity.html#identityValue\",\"classes\":\"\",\"parent\":\"types.Identity\"},{\"kind\":256,\"name\":\"Callback\",\"url\":\"interfaces/types.Callback.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":1024,\"name\":\"url\",\"url\":\"interfaces/types.Callback.html#url\",\"classes\":\"\",\"parent\":\"types.Callback\"},{\"kind\":1024,\"name\":\"headers\",\"url\":\"interfaces/types.Callback.html#headers\",\"classes\":\"\",\"parent\":\"types.Callback\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.Callback.html#headers.__type\",\"classes\":\"\",\"parent\":\"types.Callback.headers\"},{\"kind\":256,\"name\":\"Document\",\"url\":\"interfaces/types.Document.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":1024,\"name\":\"data\",\"url\":\"interfaces/types.Document.html#data\",\"classes\":\"\",\"parent\":\"types.Document\"},{\"kind\":1024,\"name\":\"url\",\"url\":\"interfaces/types.Document.html#url\",\"classes\":\"tsd-is-inherited\",\"parent\":\"types.Document\"},{\"kind\":1024,\"name\":\"headers\",\"url\":\"interfaces/types.Document.html#headers\",\"classes\":\"tsd-is-inherited\",\"parent\":\"types.Document\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.Document.html#headers.__type\",\"classes\":\"\",\"parent\":\"types.Document.headers\"},{\"kind\":256,\"name\":\"DataSubject\",\"url\":\"interfaces/types.DataSubject.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"interfaces/types.DataSubject.html#type\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"email\",\"url\":\"interfaces/types.DataSubject.html#email\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"firstName\",\"url\":\"interfaces/types.DataSubject.html#firstName\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"lastName\",\"url\":\"interfaces/types.DataSubject.html#lastName\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"addressLine1\",\"url\":\"interfaces/types.DataSubject.html#addressLine1\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"addressLine2\",\"url\":\"interfaces/types.DataSubject.html#addressLine2\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"city\",\"url\":\"interfaces/types.DataSubject.html#city\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"stateRegionCode\",\"url\":\"interfaces/types.DataSubject.html#stateRegionCode\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"postalCode\",\"url\":\"interfaces/types.DataSubject.html#postalCode\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"countryCode\",\"url\":\"interfaces/types.DataSubject.html#countryCode\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"description\",\"url\":\"interfaces/types.DataSubject.html#description\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"phone\",\"url\":\"interfaces/types.DataSubject.html#phone\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":1024,\"name\":\"formData\",\"url\":\"interfaces/types.DataSubject.html#formData\",\"classes\":\"\",\"parent\":\"types.DataSubject\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DataSubject.html#formData.__type\",\"classes\":\"\",\"parent\":\"types.DataSubject.formData\"},{\"kind\":256,\"name\":\"AccessStatusEvent\",\"url\":\"interfaces/types.AccessStatusEvent.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/types.AccessStatusEvent.html#apiVersion\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/types.AccessStatusEvent.html#kind\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/types.AccessStatusEvent.html#metadata\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.AccessStatusEvent.html#metadata.__type-3\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.metadata\"},{\"kind\":1024,\"name\":\"uid\",\"url\":\"interfaces/types.AccessStatusEvent.html#metadata.__type-3.uid\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.metadata.__type\"},{\"kind\":1024,\"name\":\"tenant\",\"url\":\"interfaces/types.AccessStatusEvent.html#metadata.__type-3.tenant\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.metadata.__type\"},{\"kind\":1024,\"name\":\"event\",\"url\":\"interfaces/types.AccessStatusEvent.html#event\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event\"},{\"kind\":1024,\"name\":\"status\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.status\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"reason\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.reason\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"expectedCompletionTimestamp\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.expectedCompletionTimestamp\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"redirectUrl\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.redirectUrl\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"requestId\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.requestId\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"documents\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.documents\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"context\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.context\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.context.__type-1\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type.context\"},{\"kind\":1024,\"name\":\"subject\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.subject\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"identities\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.identities\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"messages\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.messages\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.AccessStatusEvent.html#event.__type.messages.__type-2\",\"classes\":\"\",\"parent\":\"types.AccessStatusEvent.event.__type.messages\"},{\"kind\":256,\"name\":\"AccessRequest\",\"url\":\"interfaces/types.AccessRequest.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/types.AccessRequest.html#apiVersion\",\"classes\":\"\",\"parent\":\"types.AccessRequest\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/types.AccessRequest.html#kind\",\"classes\":\"\",\"parent\":\"types.AccessRequest\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/types.AccessRequest.html#metadata\",\"classes\":\"\",\"parent\":\"types.AccessRequest\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.AccessRequest.html#metadata.__type\",\"classes\":\"\",\"parent\":\"types.AccessRequest.metadata\"},{\"kind\":1024,\"name\":\"uid\",\"url\":\"interfaces/types.AccessRequest.html#metadata.__type.uid\",\"classes\":\"\",\"parent\":\"types.AccessRequest.metadata.__type\"},{\"kind\":1024,\"name\":\"tenant\",\"url\":\"interfaces/types.AccessRequest.html#metadata.__type.tenant\",\"classes\":\"\",\"parent\":\"types.AccessRequest.metadata.__type\"},{\"kind\":1024,\"name\":\"requestBody\",\"url\":\"interfaces/types.AccessRequest.html#requestBody\",\"classes\":\"\",\"parent\":\"types.AccessRequest\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody\"},{\"kind\":1024,\"name\":\"controller\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.controller\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"property\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.property\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"environment\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.environment\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"regulation\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.regulation\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"jurisdiction\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.jurisdiction\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"identities\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.identities\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"callbacks\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.callbacks\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"subject\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.subject\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"context\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.context\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.context.__type-2\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type.context\"},{\"kind\":1024,\"name\":\"submittedTimestamp\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.submittedTimestamp\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"dueTimestamp\",\"url\":\"interfaces/types.AccessRequest.html#requestBody.__type-1.dueTimestamp\",\"classes\":\"\",\"parent\":\"types.AccessRequest.requestBody.__type\"},{\"kind\":256,\"name\":\"AccessResponse\",\"url\":\"interfaces/types.AccessResponse.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/types.AccessResponse.html#apiVersion\",\"classes\":\"\",\"parent\":\"types.AccessResponse\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/types.AccessResponse.html#kind\",\"classes\":\"\",\"parent\":\"types.AccessResponse\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/types.AccessResponse.html#metadata\",\"classes\":\"\",\"parent\":\"types.AccessResponse\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.AccessResponse.html#metadata.__type\",\"classes\":\"\",\"parent\":\"types.AccessResponse.metadata\"},{\"kind\":1024,\"name\":\"uid\",\"url\":\"interfaces/types.AccessResponse.html#metadata.__type.uid\",\"classes\":\"\",\"parent\":\"types.AccessResponse.metadata.__type\"},{\"kind\":1024,\"name\":\"tenant\",\"url\":\"interfaces/types.AccessResponse.html#metadata.__type.tenant\",\"classes\":\"\",\"parent\":\"types.AccessResponse.metadata.__type\"},{\"kind\":1024,\"name\":\"responseBody\",\"url\":\"interfaces/types.AccessResponse.html#responseBody\",\"classes\":\"\",\"parent\":\"types.AccessResponse\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody\"},{\"kind\":1024,\"name\":\"status\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.status\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"reason\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.reason\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"expectedCompletionTimestamp\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.expectedCompletionTimestamp\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"redirectUrl\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.redirectUrl\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"requestId\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.requestId\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"documents\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.documents\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"context\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.context\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.context.__type-2\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type.context\"},{\"kind\":1024,\"name\":\"subject\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.subject\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"identities\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.identities\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"messages\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.messages\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.AccessResponse.html#responseBody.__type-1.messages.__type-3\",\"classes\":\"\",\"parent\":\"types.AccessResponse.responseBody.__type.messages\"},{\"kind\":256,\"name\":\"DeleteStatusEvent\",\"url\":\"interfaces/types.DeleteStatusEvent.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/types.DeleteStatusEvent.html#apiVersion\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/types.DeleteStatusEvent.html#kind\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/types.DeleteStatusEvent.html#metadata\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DeleteStatusEvent.html#metadata.__type-3\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.metadata\"},{\"kind\":1024,\"name\":\"uid\",\"url\":\"interfaces/types.DeleteStatusEvent.html#metadata.__type-3.uid\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.metadata.__type\"},{\"kind\":1024,\"name\":\"tenant\",\"url\":\"interfaces/types.DeleteStatusEvent.html#metadata.__type-3.tenant\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.metadata.__type\"},{\"kind\":1024,\"name\":\"event\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event\"},{\"kind\":1024,\"name\":\"status\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.status\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"reason\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.reason\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"expectedCompletionTimestamp\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.expectedCompletionTimestamp\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"redirectUrl\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.redirectUrl\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"requestId\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.requestId\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"documents\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.documents\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"context\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.context\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.context.__type-1\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type.context\"},{\"kind\":1024,\"name\":\"subject\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.subject\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"identities\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.identities\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type\"},{\"kind\":1024,\"name\":\"messages\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.messages\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DeleteStatusEvent.html#event.__type.messages.__type-2\",\"classes\":\"\",\"parent\":\"types.DeleteStatusEvent.event.__type.messages\"},{\"kind\":256,\"name\":\"DeleteRequest\",\"url\":\"interfaces/types.DeleteRequest.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/types.DeleteRequest.html#apiVersion\",\"classes\":\"\",\"parent\":\"types.DeleteRequest\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/types.DeleteRequest.html#kind\",\"classes\":\"\",\"parent\":\"types.DeleteRequest\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/types.DeleteRequest.html#metadata\",\"classes\":\"\",\"parent\":\"types.DeleteRequest\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DeleteRequest.html#metadata.__type\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.metadata\"},{\"kind\":1024,\"name\":\"uid\",\"url\":\"interfaces/types.DeleteRequest.html#metadata.__type.uid\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.metadata.__type\"},{\"kind\":1024,\"name\":\"tenant\",\"url\":\"interfaces/types.DeleteRequest.html#metadata.__type.tenant\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.metadata.__type\"},{\"kind\":1024,\"name\":\"requestBody\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody\",\"classes\":\"\",\"parent\":\"types.DeleteRequest\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody\"},{\"kind\":1024,\"name\":\"controller\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.controller\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"property\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.property\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"environment\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.environment\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"regulation\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.regulation\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"jurisdiction\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.jurisdiction\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"identities\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.identities\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"callbacks\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.callbacks\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"subject\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.subject\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"context\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.context\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.context.__type-2\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type.context\"},{\"kind\":1024,\"name\":\"submittedTimestamp\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.submittedTimestamp\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"dueTimestamp\",\"url\":\"interfaces/types.DeleteRequest.html#requestBody.__type-1.dueTimestamp\",\"classes\":\"\",\"parent\":\"types.DeleteRequest.requestBody.__type\"},{\"kind\":256,\"name\":\"DeleteResponse\",\"url\":\"interfaces/types.DeleteResponse.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/types.DeleteResponse.html#apiVersion\",\"classes\":\"\",\"parent\":\"types.DeleteResponse\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/types.DeleteResponse.html#kind\",\"classes\":\"\",\"parent\":\"types.DeleteResponse\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/types.DeleteResponse.html#metadata\",\"classes\":\"\",\"parent\":\"types.DeleteResponse\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DeleteResponse.html#metadata.__type\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.metadata\"},{\"kind\":1024,\"name\":\"uid\",\"url\":\"interfaces/types.DeleteResponse.html#metadata.__type.uid\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.metadata.__type\"},{\"kind\":1024,\"name\":\"tenant\",\"url\":\"interfaces/types.DeleteResponse.html#metadata.__type.tenant\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.metadata.__type\"},{\"kind\":1024,\"name\":\"responseBody\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody\",\"classes\":\"\",\"parent\":\"types.DeleteResponse\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody\"},{\"kind\":1024,\"name\":\"status\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.status\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"reason\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.reason\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"expectedCompletionTimestamp\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.expectedCompletionTimestamp\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"redirectUrl\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.redirectUrl\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"requestId\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.requestId\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"documents\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.documents\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"context\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.context\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.context.__type-2\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type.context\"},{\"kind\":1024,\"name\":\"subject\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.subject\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"identities\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.identities\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type\"},{\"kind\":1024,\"name\":\"messages\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.messages\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/types.DeleteResponse.html#responseBody.__type-1.messages.__type-3\",\"classes\":\"\",\"parent\":\"types.DeleteResponse.responseBody.__type.messages\"},{\"kind\":4194304,\"name\":\"DsrRequest\",\"url\":\"types/types.DsrRequest.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":4194304,\"name\":\"DsrResponse\",\"url\":\"types/types.DsrResponse.html\",\"classes\":\"\",\"parent\":\"types\"},{\"kind\":256,\"name\":\"ConnectionConfig\",\"url\":\"interfaces/types.ConnectionConfig.html\",\"classes\":\"\",\"parent\":\"types\"}],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[[\"name/0\",[0,48.31]],[\"comment/0\",[]],[\"name/1\",[1,48.31]],[\"comment/1\",[]],[\"name/2\",[2,48.31]],[\"comment/2\",[]],[\"name/3\",[3,48.31]],[\"comment/3\",[]],[\"name/4\",[4,48.31]],[\"comment/4\",[]],[\"name/5\",[5,48.31]],[\"comment/5\",[]],[\"name/6\",[6,48.31]],[\"comment/6\",[]],[\"name/7\",[7,48.31]],[\"comment/7\",[]],[\"name/8\",[8,48.31]],[\"comment/8\",[]],[\"name/9\",[9,48.31]],[\"comment/9\",[]],[\"name/10\",[10,48.31]],[\"comment/10\",[]],[\"name/11\",[11,48.31]],[\"comment/11\",[]],[\"name/12\",[12,48.31]],[\"comment/12\",[]],[\"name/13\",[13,48.31]],[\"comment/13\",[]],[\"name/14\",[14,48.31]],[\"comment/14\",[]],[\"name/15\",[15,48.31]],[\"comment/15\",[]],[\"name/16\",[16,48.31]],[\"comment/16\",[]],[\"name/17\",[17,48.31]],[\"comment/17\",[]],[\"name/18\",[18,48.31]],[\"comment/18\",[]],[\"name/19\",[19,48.31]],[\"comment/19\",[]],[\"name/20\",[20,48.31]],[\"comment/20\",[]],[\"name/21\",[21,48.31]],[\"comment/21\",[]],[\"name/22\",[22,48.31]],[\"comment/22\",[]],[\"name/23\",[23,32.215]],[\"comment/23\",[]],[\"name/24\",[24,43.202]],[\"comment/24\",[]],[\"name/25\",[25,43.202]],[\"comment/25\",[]],[\"name/26\",[26,43.202]],[\"comment/26\",[]],[\"name/27\",[27,43.202]],[\"comment/27\",[]],[\"name/28\",[28,43.202]],[\"comment/28\",[]],[\"name/29\",[29,43.202]],[\"comment/29\",[]],[\"name/30\",[30,48.31]],[\"comment/30\",[]],[\"name/31\",[31,48.31]],[\"comment/31\",[]],[\"name/32\",[32,48.31]],[\"comment/32\",[]],[\"name/33\",[33,48.31]],[\"comment/33\",[]],[\"name/34\",[34,48.31]],[\"comment/34\",[]],[\"name/35\",[35,43.202]],[\"comment/35\",[]],[\"name/36\",[36,43.202]],[\"comment/36\",[]],[\"name/37\",[37,19.978]],[\"comment/37\",[]],[\"name/38\",[38,48.31]],[\"comment/38\",[]],[\"name/39\",[39,48.31]],[\"comment/39\",[]],[\"name/40\",[35,43.202]],[\"comment/40\",[]],[\"name/41\",[36,43.202]],[\"comment/41\",[]],[\"name/42\",[37,19.978]],[\"comment/42\",[]],[\"name/43\",[40,48.31]],[\"comment/43\",[]],[\"name/44\",[41,48.31]],[\"comment/44\",[]],[\"name/45\",[42,48.31]],[\"comment/45\",[]],[\"name/46\",[43,48.31]],[\"comment/46\",[]],[\"name/47\",[44,48.31]],[\"comment/47\",[]],[\"name/48\",[45,48.31]],[\"comment/48\",[]],[\"name/49\",[46,48.31]],[\"comment/49\",[]],[\"name/50\",[47,48.31]],[\"comment/50\",[]],[\"name/51\",[48,48.31]],[\"comment/51\",[]],[\"name/52\",[49,48.31]],[\"comment/52\",[]],[\"name/53\",[50,48.31]],[\"comment/53\",[]],[\"name/54\",[51,48.31]],[\"comment/54\",[]],[\"name/55\",[52,48.31]],[\"comment/55\",[]],[\"name/56\",[53,48.31]],[\"comment/56\",[]],[\"name/57\",[37,19.978]],[\"comment/57\",[]],[\"name/58\",[26,43.202]],[\"comment/58\",[]],[\"name/59\",[54,33.646]],[\"comment/59\",[]],[\"name/60\",[23,32.215]],[\"comment/60\",[]],[\"name/61\",[55,33.646]],[\"comment/61\",[]],[\"name/62\",[37,19.978]],[\"comment/62\",[]],[\"name/63\",[56,33.646]],[\"comment/63\",[]],[\"name/64\",[57,33.646]],[\"comment/64\",[]],[\"name/65\",[58,43.202]],[\"comment/65\",[]],[\"name/66\",[37,19.978]],[\"comment/66\",[]],[\"name/67\",[59,37.324]],[\"comment/67\",[]],[\"name/68\",[60,37.324]],[\"comment/68\",[]],[\"name/69\",[61,37.324]],[\"comment/69\",[]],[\"name/70\",[62,37.324]],[\"comment/70\",[]],[\"name/71\",[63,37.324]],[\"comment/71\",[]],[\"name/72\",[64,37.324]],[\"comment/72\",[]],[\"name/73\",[65,33.646]],[\"comment/73\",[]],[\"name/74\",[37,19.978]],[\"comment/74\",[]],[\"name/75\",[66,33.646]],[\"comment/75\",[]],[\"name/76\",[67,33.646]],[\"comment/76\",[]],[\"name/77\",[68,37.324]],[\"comment/77\",[]],[\"name/78\",[37,19.978]],[\"comment/78\",[]],[\"name/79\",[24,43.202]],[\"comment/79\",[]],[\"name/80\",[54,33.646]],[\"comment/80\",[]],[\"name/81\",[23,32.215]],[\"comment/81\",[]],[\"name/82\",[55,33.646]],[\"comment/82\",[]],[\"name/83\",[37,19.978]],[\"comment/83\",[]],[\"name/84\",[56,33.646]],[\"comment/84\",[]],[\"name/85\",[57,33.646]],[\"comment/85\",[]],[\"name/86\",[69,43.202]],[\"comment/86\",[]],[\"name/87\",[37,19.978]],[\"comment/87\",[]],[\"name/88\",[70,43.202]],[\"comment/88\",[]],[\"name/89\",[71,43.202]],[\"comment/89\",[]],[\"name/90\",[72,43.202]],[\"comment/90\",[]],[\"name/91\",[73,43.202]],[\"comment/91\",[]],[\"name/92\",[74,43.202]],[\"comment/92\",[]],[\"name/93\",[67,33.646]],[\"comment/93\",[]],[\"name/94\",[75,43.202]],[\"comment/94\",[]],[\"name/95\",[66,33.646]],[\"comment/95\",[]],[\"name/96\",[65,33.646]],[\"comment/96\",[]],[\"name/97\",[37,19.978]],[\"comment/97\",[]],[\"name/98\",[76,43.202]],[\"comment/98\",[]],[\"name/99\",[77,43.202]],[\"comment/99\",[]],[\"name/100\",[25,43.202]],[\"comment/100\",[]],[\"name/101\",[54,33.646]],[\"comment/101\",[]],[\"name/102\",[23,32.215]],[\"comment/102\",[]],[\"name/103\",[55,33.646]],[\"comment/103\",[]],[\"name/104\",[37,19.978]],[\"comment/104\",[]],[\"name/105\",[56,33.646]],[\"comment/105\",[]],[\"name/106\",[57,33.646]],[\"comment/106\",[]],[\"name/107\",[78,43.202]],[\"comment/107\",[]],[\"name/108\",[37,19.978]],[\"comment/108\",[]],[\"name/109\",[59,37.324]],[\"comment/109\",[]],[\"name/110\",[60,37.324]],[\"comment/110\",[]],[\"name/111\",[61,37.324]],[\"comment/111\",[]],[\"name/112\",[62,37.324]],[\"comment/112\",[]],[\"name/113\",[63,37.324]],[\"comment/113\",[]],[\"name/114\",[64,37.324]],[\"comment/114\",[]],[\"name/115\",[65,33.646]],[\"comment/115\",[]],[\"name/116\",[37,19.978]],[\"comment/116\",[]],[\"name/117\",[66,33.646]],[\"comment/117\",[]],[\"name/118\",[67,33.646]],[\"comment/118\",[]],[\"name/119\",[68,37.324]],[\"comment/119\",[]],[\"name/120\",[37,19.978]],[\"comment/120\",[]],[\"name/121\",[29,43.202]],[\"comment/121\",[]],[\"name/122\",[54,33.646]],[\"comment/122\",[]],[\"name/123\",[23,32.215]],[\"comment/123\",[]],[\"name/124\",[55,33.646]],[\"comment/124\",[]],[\"name/125\",[37,19.978]],[\"comment/125\",[]],[\"name/126\",[56,33.646]],[\"comment/126\",[]],[\"name/127\",[57,33.646]],[\"comment/127\",[]],[\"name/128\",[58,43.202]],[\"comment/128\",[]],[\"name/129\",[37,19.978]],[\"comment/129\",[]],[\"name/130\",[59,37.324]],[\"comment/130\",[]],[\"name/131\",[60,37.324]],[\"comment/131\",[]],[\"name/132\",[61,37.324]],[\"comment/132\",[]],[\"name/133\",[62,37.324]],[\"comment/133\",[]],[\"name/134\",[63,37.324]],[\"comment/134\",[]],[\"name/135\",[64,37.324]],[\"comment/135\",[]],[\"name/136\",[65,33.646]],[\"comment/136\",[]],[\"name/137\",[37,19.978]],[\"comment/137\",[]],[\"name/138\",[66,33.646]],[\"comment/138\",[]],[\"name/139\",[67,33.646]],[\"comment/139\",[]],[\"name/140\",[68,37.324]],[\"comment/140\",[]],[\"name/141\",[37,19.978]],[\"comment/141\",[]],[\"name/142\",[27,43.202]],[\"comment/142\",[]],[\"name/143\",[54,33.646]],[\"comment/143\",[]],[\"name/144\",[23,32.215]],[\"comment/144\",[]],[\"name/145\",[55,33.646]],[\"comment/145\",[]],[\"name/146\",[37,19.978]],[\"comment/146\",[]],[\"name/147\",[56,33.646]],[\"comment/147\",[]],[\"name/148\",[57,33.646]],[\"comment/148\",[]],[\"name/149\",[69,43.202]],[\"comment/149\",[]],[\"name/150\",[37,19.978]],[\"comment/150\",[]],[\"name/151\",[70,43.202]],[\"comment/151\",[]],[\"name/152\",[71,43.202]],[\"comment/152\",[]],[\"name/153\",[72,43.202]],[\"comment/153\",[]],[\"name/154\",[73,43.202]],[\"comment/154\",[]],[\"name/155\",[74,43.202]],[\"comment/155\",[]],[\"name/156\",[67,33.646]],[\"comment/156\",[]],[\"name/157\",[75,43.202]],[\"comment/157\",[]],[\"name/158\",[66,33.646]],[\"comment/158\",[]],[\"name/159\",[65,33.646]],[\"comment/159\",[]],[\"name/160\",[37,19.978]],[\"comment/160\",[]],[\"name/161\",[76,43.202]],[\"comment/161\",[]],[\"name/162\",[77,43.202]],[\"comment/162\",[]],[\"name/163\",[28,43.202]],[\"comment/163\",[]],[\"name/164\",[54,33.646]],[\"comment/164\",[]],[\"name/165\",[23,32.215]],[\"comment/165\",[]],[\"name/166\",[55,33.646]],[\"comment/166\",[]],[\"name/167\",[37,19.978]],[\"comment/167\",[]],[\"name/168\",[56,33.646]],[\"comment/168\",[]],[\"name/169\",[57,33.646]],[\"comment/169\",[]],[\"name/170\",[78,43.202]],[\"comment/170\",[]],[\"name/171\",[37,19.978]],[\"comment/171\",[]],[\"name/172\",[59,37.324]],[\"comment/172\",[]],[\"name/173\",[60,37.324]],[\"comment/173\",[]],[\"name/174\",[61,37.324]],[\"comment/174\",[]],[\"name/175\",[62,37.324]],[\"comment/175\",[]],[\"name/176\",[63,37.324]],[\"comment/176\",[]],[\"name/177\",[64,37.324]],[\"comment/177\",[]],[\"name/178\",[65,33.646]],[\"comment/178\",[]],[\"name/179\",[37,19.978]],[\"comment/179\",[]],[\"name/180\",[66,33.646]],[\"comment/180\",[]],[\"name/181\",[67,33.646]],[\"comment/181\",[]],[\"name/182\",[68,37.324]],[\"comment/182\",[]],[\"name/183\",[37,19.978]],[\"comment/183\",[]],[\"name/184\",[79,48.31]],[\"comment/184\",[]],[\"name/185\",[80,48.31]],[\"comment/185\",[]],[\"name/186\",[81,48.31]],[\"comment/186\",[]]],\"invertedIndex\":[[\"__type\",{\"_index\":37,\"name\":{\"37\":{},\"42\":{},\"57\":{},\"62\":{},\"66\":{},\"74\":{},\"78\":{},\"83\":{},\"87\":{},\"97\":{},\"104\":{},\"108\":{},\"116\":{},\"120\":{},\"125\":{},\"129\":{},\"137\":{},\"141\":{},\"146\":{},\"150\":{},\"160\":{},\"167\":{},\"171\":{},\"179\":{},\"183\":{}},\"comment\":{}}],[\"accessrequest\",{\"_index\":24,\"name\":{\"24\":{},\"79\":{}},\"comment\":{}}],[\"accessresponse\",{\"_index\":25,\"name\":{\"25\":{},\"100\":{}},\"comment\":{}}],[\"accessstatusevent\",{\"_index\":26,\"name\":{\"26\":{},\"58\":{}},\"comment\":{}}],[\"addressline1\",{\"_index\":45,\"name\":{\"48\":{}},\"comment\":{}}],[\"addressline2\",{\"_index\":46,\"name\":{\"49\":{}},\"comment\":{}}],[\"apiversion\",{\"_index\":54,\"name\":{\"59\":{},\"80\":{},\"101\":{},\"122\":{},\"143\":{},\"164\":{}},\"comment\":{}}],[\"callback\",{\"_index\":34,\"name\":{\"34\":{}},\"comment\":{}}],[\"callbacks\",{\"_index\":75,\"name\":{\"94\":{},\"157\":{}},\"comment\":{}}],[\"cancelledrequeststatus\",{\"_index\":8,\"name\":{\"8\":{}},\"comment\":{}}],[\"city\",{\"_index\":47,\"name\":{\"50\":{}},\"comment\":{}}],[\"claimnotcoveredrequeststatusreason\",{\"_index\":19,\"name\":{\"19\":{}},\"comment\":{}}],[\"completedrequeststatus\",{\"_index\":7,\"name\":{\"7\":{}},\"comment\":{}}],[\"connectionconfig\",{\"_index\":81,\"name\":{\"186\":{}},\"comment\":{}}],[\"context\",{\"_index\":65,\"name\":{\"73\":{},\"96\":{},\"115\":{},\"136\":{},\"159\":{},\"178\":{}},\"comment\":{}}],[\"controller\",{\"_index\":70,\"name\":{\"88\":{},\"151\":{}},\"comment\":{}}],[\"countrycode\",{\"_index\":50,\"name\":{\"53\":{}},\"comment\":{}}],[\"data\",{\"_index\":39,\"name\":{\"39\":{}},\"comment\":{}}],[\"datasubject\",{\"_index\":40,\"name\":{\"43\":{}},\"comment\":{}}],[\"deleterequest\",{\"_index\":27,\"name\":{\"27\":{},\"142\":{}},\"comment\":{}}],[\"deleteresponse\",{\"_index\":28,\"name\":{\"28\":{},\"163\":{}},\"comment\":{}}],[\"deletestatusevent\",{\"_index\":29,\"name\":{\"29\":{},\"121\":{}},\"comment\":{}}],[\"deniedrequeststatus\",{\"_index\":9,\"name\":{\"9\":{}},\"comment\":{}}],[\"description\",{\"_index\":51,\"name\":{\"54\":{}},\"comment\":{}}],[\"document\",{\"_index\":38,\"name\":{\"38\":{}},\"comment\":{}}],[\"documents\",{\"_index\":64,\"name\":{\"72\":{},\"114\":{},\"135\":{},\"177\":{}},\"comment\":{}}],[\"dsrrequest\",{\"_index\":79,\"name\":{\"184\":{}},\"comment\":{}}],[\"dsrresponse\",{\"_index\":80,\"name\":{\"185\":{}},\"comment\":{}}],[\"duetimestamp\",{\"_index\":77,\"name\":{\"99\":{},\"162\":{}},\"comment\":{}}],[\"email\",{\"_index\":42,\"name\":{\"45\":{}},\"comment\":{}}],[\"environment\",{\"_index\":72,\"name\":{\"90\":{},\"153\":{}},\"comment\":{}}],[\"event\",{\"_index\":58,\"name\":{\"65\":{},\"128\":{}},\"comment\":{}}],[\"executed\",{\"_index\":15,\"name\":{\"15\":{}},\"comment\":{}}],[\"expectedcompletiontimestamp\",{\"_index\":61,\"name\":{\"69\":{},\"111\":{},\"132\":{},\"174\":{}},\"comment\":{}}],[\"firstname\",{\"_index\":43,\"name\":{\"46\":{}},\"comment\":{}}],[\"formdata\",{\"_index\":53,\"name\":{\"56\":{}},\"comment\":{}}],[\"handlerequest\",{\"_index\":1,\"name\":{\"1\":{}},\"comment\":{}}],[\"headers\",{\"_index\":36,\"name\":{\"36\":{},\"41\":{}},\"comment\":{}}],[\"identities\",{\"_index\":67,\"name\":{\"76\":{},\"93\":{},\"118\":{},\"139\":{},\"156\":{},\"181\":{}},\"comment\":{}}],[\"identity\",{\"_index\":30,\"name\":{\"30\":{}},\"comment\":{}}],[\"identityformat\",{\"_index\":32,\"name\":{\"32\":{}},\"comment\":{}}],[\"identityspace\",{\"_index\":31,\"name\":{\"31\":{}},\"comment\":{}}],[\"identityvalue\",{\"_index\":33,\"name\":{\"33\":{}},\"comment\":{}}],[\"index\",{\"_index\":0,\"name\":{\"0\":{}},\"comment\":{}}],[\"inprogressrequeststatus\",{\"_index\":6,\"name\":{\"6\":{}},\"comment\":{}}],[\"insufficientidentification\",{\"_index\":14,\"name\":{\"14\":{}},\"comment\":{}}],[\"insufficientverificationrequeststatusreason\",{\"_index\":17,\"name\":{\"17\":{}},\"comment\":{}}],[\"jurisdiction\",{\"_index\":74,\"name\":{\"92\":{},\"155\":{}},\"comment\":{}}],[\"kind\",{\"_index\":23,\"name\":{\"23\":{},\"60\":{},\"81\":{},\"102\":{},\"123\":{},\"144\":{},\"165\":{}},\"comment\":{}}],[\"lastname\",{\"_index\":44,\"name\":{\"47\":{}},\"comment\":{}}],[\"messages\",{\"_index\":68,\"name\":{\"77\":{},\"119\":{},\"140\":{},\"182\":{}},\"comment\":{}}],[\"metadata\",{\"_index\":55,\"name\":{\"61\":{},\"82\":{},\"103\":{},\"124\":{},\"145\":{},\"166\":{}},\"comment\":{}}],[\"needuserverificationrequeststatusreason\",{\"_index\":12,\"name\":{\"12\":{}},\"comment\":{}}],[\"nomatchrequeststatusreason\",{\"_index\":18,\"name\":{\"18\":{}},\"comment\":{}}],[\"otherrequeststatusreason\",{\"_index\":22,\"name\":{\"22\":{}},\"comment\":{}}],[\"outsidejurisdictionrequeststatusreason\",{\"_index\":20,\"name\":{\"20\":{}},\"comment\":{}}],[\"pendingrequeststatus\",{\"_index\":5,\"name\":{\"5\":{}},\"comment\":{}}],[\"phone\",{\"_index\":52,\"name\":{\"55\":{}},\"comment\":{}}],[\"postalcode\",{\"_index\":49,\"name\":{\"52\":{}},\"comment\":{}}],[\"property\",{\"_index\":71,\"name\":{\"89\":{},\"152\":{}},\"comment\":{}}],[\"reason\",{\"_index\":60,\"name\":{\"68\":{},\"110\":{},\"131\":{},\"173\":{}},\"comment\":{}}],[\"redirecturl\",{\"_index\":62,\"name\":{\"70\":{},\"112\":{},\"133\":{},\"175\":{}},\"comment\":{}}],[\"regulation\",{\"_index\":73,\"name\":{\"91\":{},\"154\":{}},\"comment\":{}}],[\"requestbody\",{\"_index\":69,\"name\":{\"86\":{},\"149\":{}},\"comment\":{}}],[\"requested\",{\"_index\":13,\"name\":{\"13\":{}},\"comment\":{}}],[\"requestid\",{\"_index\":63,\"name\":{\"71\":{},\"113\":{},\"134\":{},\"176\":{}},\"comment\":{}}],[\"requeststatus\",{\"_index\":3,\"name\":{\"3\":{}},\"comment\":{}}],[\"requeststatusreason\",{\"_index\":10,\"name\":{\"10\":{}},\"comment\":{}}],[\"responsebody\",{\"_index\":78,\"name\":{\"107\":{},\"170\":{}},\"comment\":{}}],[\"stateregioncode\",{\"_index\":48,\"name\":{\"51\":{}},\"comment\":{}}],[\"status\",{\"_index\":59,\"name\":{\"67\":{},\"109\":{},\"130\":{},\"172\":{}},\"comment\":{}}],[\"subject\",{\"_index\":66,\"name\":{\"75\":{},\"95\":{},\"117\":{},\"138\":{},\"158\":{},\"180\":{}},\"comment\":{}}],[\"submittedtimestamp\",{\"_index\":76,\"name\":{\"98\":{},\"161\":{}},\"comment\":{}}],[\"suspectedfraudrequeststatusreason\",{\"_index\":16,\"name\":{\"16\":{}},\"comment\":{}}],[\"tenant\",{\"_index\":57,\"name\":{\"64\":{},\"85\":{},\"106\":{},\"127\":{},\"148\":{},\"169\":{}},\"comment\":{}}],[\"toomanyrequestsrequeststatusreason\",{\"_index\":21,\"name\":{\"21\":{}},\"comment\":{}}],[\"type\",{\"_index\":41,\"name\":{\"44\":{}},\"comment\":{}}],[\"types\",{\"_index\":2,\"name\":{\"2\":{}},\"comment\":{}}],[\"uid\",{\"_index\":56,\"name\":{\"63\":{},\"84\":{},\"105\":{},\"126\":{},\"147\":{},\"168\":{}},\"comment\":{}}],[\"unknownrequeststatus\",{\"_index\":4,\"name\":{\"4\":{}},\"comment\":{}}],[\"unknownrequeststatusreason\",{\"_index\":11,\"name\":{\"11\":{}},\"comment\":{}}],[\"url\",{\"_index\":35,\"name\":{\"35\":{},\"40\":{}},\"comment\":{}}]],\"pipeline\":[]}}");
+window.searchData = JSON.parse("{\"rows\":[{\"kind\":4194304,\"name\":\"HandleRequest\",\"url\":\"types/HandleRequest.html\",\"classes\":\"\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/HandleRequest.html#__type\",\"classes\":\"\",\"parent\":\"HandleRequest\"},{\"kind\":4194304,\"name\":\"DSRVersion\",\"url\":\"types/DSRVersion.html\",\"classes\":\"\"},{\"kind\":8,\"name\":\"RequestStatus\",\"url\":\"enums/RequestStatus.html\",\"classes\":\"\"},{\"kind\":16,\"name\":\"UnknownRequestStatus\",\"url\":\"enums/RequestStatus.html#UnknownRequestStatus\",\"classes\":\"\",\"parent\":\"RequestStatus\"},{\"kind\":16,\"name\":\"PendingRequestStatus\",\"url\":\"enums/RequestStatus.html#PendingRequestStatus\",\"classes\":\"\",\"parent\":\"RequestStatus\"},{\"kind\":16,\"name\":\"InProgressRequestStatus\",\"url\":\"enums/RequestStatus.html#InProgressRequestStatus\",\"classes\":\"\",\"parent\":\"RequestStatus\"},{\"kind\":16,\"name\":\"CompletedRequestStatus\",\"url\":\"enums/RequestStatus.html#CompletedRequestStatus\",\"classes\":\"\",\"parent\":\"RequestStatus\"},{\"kind\":16,\"name\":\"CancelledRequestStatus\",\"url\":\"enums/RequestStatus.html#CancelledRequestStatus\",\"classes\":\"\",\"parent\":\"RequestStatus\"},{\"kind\":16,\"name\":\"DeniedRequestStatus\",\"url\":\"enums/RequestStatus.html#DeniedRequestStatus\",\"classes\":\"\",\"parent\":\"RequestStatus\"},{\"kind\":8,\"name\":\"RequestStatusReason\",\"url\":\"enums/RequestStatusReason.html\",\"classes\":\"\"},{\"kind\":16,\"name\":\"UnknownRequestStatusReason\",\"url\":\"enums/RequestStatusReason.html#UnknownRequestStatusReason\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":16,\"name\":\"NeedUserVerificationRequestStatusReason\",\"url\":\"enums/RequestStatusReason.html#NeedUserVerificationRequestStatusReason\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":16,\"name\":\"Requested\",\"url\":\"enums/RequestStatusReason.html#Requested\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":16,\"name\":\"InsufficientIdentification\",\"url\":\"enums/RequestStatusReason.html#InsufficientIdentification\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":16,\"name\":\"Executed\",\"url\":\"enums/RequestStatusReason.html#Executed\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":16,\"name\":\"SuspectedFraudRequestStatusReason\",\"url\":\"enums/RequestStatusReason.html#SuspectedFraudRequestStatusReason\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":16,\"name\":\"InsufficientVerificationRequestStatusReason\",\"url\":\"enums/RequestStatusReason.html#InsufficientVerificationRequestStatusReason\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":16,\"name\":\"NoMatchRequestStatusReason\",\"url\":\"enums/RequestStatusReason.html#NoMatchRequestStatusReason\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":16,\"name\":\"ClaimNotCoveredRequestStatusReason\",\"url\":\"enums/RequestStatusReason.html#ClaimNotCoveredRequestStatusReason\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":16,\"name\":\"OutsideJurisdictionRequestStatusReason\",\"url\":\"enums/RequestStatusReason.html#OutsideJurisdictionRequestStatusReason\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":16,\"name\":\"TooManyRequestsRequestStatusReason\",\"url\":\"enums/RequestStatusReason.html#TooManyRequestsRequestStatusReason\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":16,\"name\":\"OtherRequestStatusReason\",\"url\":\"enums/RequestStatusReason.html#OtherRequestStatusReason\",\"classes\":\"\",\"parent\":\"RequestStatusReason\"},{\"kind\":8,\"name\":\"Kind\",\"url\":\"enums/Kind.html\",\"classes\":\"\"},{\"kind\":16,\"name\":\"AccessRequest\",\"url\":\"enums/Kind.html#AccessRequest\",\"classes\":\"\",\"parent\":\"Kind\"},{\"kind\":16,\"name\":\"AccessResponse\",\"url\":\"enums/Kind.html#AccessResponse\",\"classes\":\"\",\"parent\":\"Kind\"},{\"kind\":16,\"name\":\"AccessStatusEvent\",\"url\":\"enums/Kind.html#AccessStatusEvent\",\"classes\":\"\",\"parent\":\"Kind\"},{\"kind\":16,\"name\":\"DeleteRequest\",\"url\":\"enums/Kind.html#DeleteRequest\",\"classes\":\"\",\"parent\":\"Kind\"},{\"kind\":16,\"name\":\"DeleteResponse\",\"url\":\"enums/Kind.html#DeleteResponse\",\"classes\":\"\",\"parent\":\"Kind\"},{\"kind\":16,\"name\":\"DeleteStatusEvent\",\"url\":\"enums/Kind.html#DeleteStatusEvent\",\"classes\":\"\",\"parent\":\"Kind\"},{\"kind\":256,\"name\":\"Identity\",\"url\":\"interfaces/Identity.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"identitySpace\",\"url\":\"interfaces/Identity.html#identitySpace\",\"classes\":\"\",\"parent\":\"Identity\"},{\"kind\":1024,\"name\":\"identityFormat\",\"url\":\"interfaces/Identity.html#identityFormat\",\"classes\":\"\",\"parent\":\"Identity\"},{\"kind\":1024,\"name\":\"identityValue\",\"url\":\"interfaces/Identity.html#identityValue\",\"classes\":\"\",\"parent\":\"Identity\"},{\"kind\":256,\"name\":\"Callback\",\"url\":\"interfaces/Callback.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"url\",\"url\":\"interfaces/Callback.html#url\",\"classes\":\"\",\"parent\":\"Callback\"},{\"kind\":1024,\"name\":\"headers\",\"url\":\"interfaces/Callback.html#headers\",\"classes\":\"\",\"parent\":\"Callback\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/Callback.html#headers.__type\",\"classes\":\"\",\"parent\":\"Callback.headers\"},{\"kind\":256,\"name\":\"Document\",\"url\":\"interfaces/Document.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"data\",\"url\":\"interfaces/Document.html#data\",\"classes\":\"\",\"parent\":\"Document\"},{\"kind\":1024,\"name\":\"url\",\"url\":\"interfaces/Document.html#url\",\"classes\":\"tsd-is-inherited\",\"parent\":\"Document\"},{\"kind\":1024,\"name\":\"headers\",\"url\":\"interfaces/Document.html#headers\",\"classes\":\"tsd-is-inherited\",\"parent\":\"Document\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/Document.html#headers.__type\",\"classes\":\"\",\"parent\":\"Document.headers\"},{\"kind\":256,\"name\":\"DataSubject\",\"url\":\"interfaces/DataSubject.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"interfaces/DataSubject.html#type\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"email\",\"url\":\"interfaces/DataSubject.html#email\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"firstName\",\"url\":\"interfaces/DataSubject.html#firstName\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"lastName\",\"url\":\"interfaces/DataSubject.html#lastName\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"addressLine1\",\"url\":\"interfaces/DataSubject.html#addressLine1\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"addressLine2\",\"url\":\"interfaces/DataSubject.html#addressLine2\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"city\",\"url\":\"interfaces/DataSubject.html#city\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"stateRegionCode\",\"url\":\"interfaces/DataSubject.html#stateRegionCode\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"postalCode\",\"url\":\"interfaces/DataSubject.html#postalCode\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"countryCode\",\"url\":\"interfaces/DataSubject.html#countryCode\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"description\",\"url\":\"interfaces/DataSubject.html#description\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"phone\",\"url\":\"interfaces/DataSubject.html#phone\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":1024,\"name\":\"formData\",\"url\":\"interfaces/DataSubject.html#formData\",\"classes\":\"\",\"parent\":\"DataSubject\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/DataSubject.html#formData.__type\",\"classes\":\"\",\"parent\":\"DataSubject.formData\"},{\"kind\":256,\"name\":\"Metadata\",\"url\":\"interfaces/Metadata.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"uid\",\"url\":\"interfaces/Metadata.html#uid\",\"classes\":\"\",\"parent\":\"Metadata\"},{\"kind\":1024,\"name\":\"tenant\",\"url\":\"interfaces/Metadata.html#tenant\",\"classes\":\"\",\"parent\":\"Metadata\"},{\"kind\":256,\"name\":\"AccessRequest\",\"url\":\"interfaces/AccessRequest.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/AccessRequest.html#apiVersion\",\"classes\":\"\",\"parent\":\"AccessRequest\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/AccessRequest.html#kind\",\"classes\":\"\",\"parent\":\"AccessRequest\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/AccessRequest.html#metadata\",\"classes\":\"\",\"parent\":\"AccessRequest\"},{\"kind\":1024,\"name\":\"requestBody\",\"url\":\"interfaces/AccessRequest.html#requestBody\",\"classes\":\"\",\"parent\":\"AccessRequest\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody\"},{\"kind\":1024,\"name\":\"controller\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.controller\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"property\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.property\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"environment\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.environment\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"regulation\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.regulation\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"jurisdiction\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.jurisdiction\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"identities\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.identities\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"callbacks\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.callbacks\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"subject\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.subject\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"context\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.context\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.context.__type-1\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type.context\"},{\"kind\":1024,\"name\":\"submittedTimestamp\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.submittedTimestamp\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"dueTimestamp\",\"url\":\"interfaces/AccessRequest.html#requestBody.__type.dueTimestamp\",\"classes\":\"\",\"parent\":\"AccessRequest.requestBody.__type\"},{\"kind\":256,\"name\":\"AccessResponseBody\",\"url\":\"interfaces/AccessResponseBody.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"status\",\"url\":\"interfaces/AccessResponseBody.html#status\",\"classes\":\"\",\"parent\":\"AccessResponseBody\"},{\"kind\":1024,\"name\":\"reason\",\"url\":\"interfaces/AccessResponseBody.html#reason\",\"classes\":\"\",\"parent\":\"AccessResponseBody\"},{\"kind\":1024,\"name\":\"expectedCompletionTimestamp\",\"url\":\"interfaces/AccessResponseBody.html#expectedCompletionTimestamp\",\"classes\":\"\",\"parent\":\"AccessResponseBody\"},{\"kind\":1024,\"name\":\"redirectUrl\",\"url\":\"interfaces/AccessResponseBody.html#redirectUrl\",\"classes\":\"\",\"parent\":\"AccessResponseBody\"},{\"kind\":1024,\"name\":\"requestId\",\"url\":\"interfaces/AccessResponseBody.html#requestId\",\"classes\":\"\",\"parent\":\"AccessResponseBody\"},{\"kind\":1024,\"name\":\"documents\",\"url\":\"interfaces/AccessResponseBody.html#documents\",\"classes\":\"\",\"parent\":\"AccessResponseBody\"},{\"kind\":1024,\"name\":\"context\",\"url\":\"interfaces/AccessResponseBody.html#context\",\"classes\":\"\",\"parent\":\"AccessResponseBody\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/AccessResponseBody.html#context.__type\",\"classes\":\"\",\"parent\":\"AccessResponseBody.context\"},{\"kind\":1024,\"name\":\"subject\",\"url\":\"interfaces/AccessResponseBody.html#subject\",\"classes\":\"\",\"parent\":\"AccessResponseBody\"},{\"kind\":1024,\"name\":\"identities\",\"url\":\"interfaces/AccessResponseBody.html#identities\",\"classes\":\"\",\"parent\":\"AccessResponseBody\"},{\"kind\":256,\"name\":\"AccessResponse\",\"url\":\"interfaces/AccessResponse.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/AccessResponse.html#apiVersion\",\"classes\":\"\",\"parent\":\"AccessResponse\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/AccessResponse.html#kind\",\"classes\":\"\",\"parent\":\"AccessResponse\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/AccessResponse.html#metadata\",\"classes\":\"\",\"parent\":\"AccessResponse\"},{\"kind\":1024,\"name\":\"responseBody\",\"url\":\"interfaces/AccessResponse.html#responseBody\",\"classes\":\"\",\"parent\":\"AccessResponse\"},{\"kind\":256,\"name\":\"AccessStatusEvent\",\"url\":\"interfaces/AccessStatusEvent.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/AccessStatusEvent.html#apiVersion\",\"classes\":\"\",\"parent\":\"AccessStatusEvent\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/AccessStatusEvent.html#kind\",\"classes\":\"\",\"parent\":\"AccessStatusEvent\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/AccessStatusEvent.html#metadata\",\"classes\":\"\",\"parent\":\"AccessStatusEvent\"},{\"kind\":1024,\"name\":\"event\",\"url\":\"interfaces/AccessStatusEvent.html#event\",\"classes\":\"\",\"parent\":\"AccessStatusEvent\"},{\"kind\":256,\"name\":\"DeleteRequest\",\"url\":\"interfaces/DeleteRequest.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/DeleteRequest.html#apiVersion\",\"classes\":\"\",\"parent\":\"DeleteRequest\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/DeleteRequest.html#kind\",\"classes\":\"\",\"parent\":\"DeleteRequest\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/DeleteRequest.html#metadata\",\"classes\":\"\",\"parent\":\"DeleteRequest\"},{\"kind\":1024,\"name\":\"requestBody\",\"url\":\"interfaces/DeleteRequest.html#requestBody\",\"classes\":\"\",\"parent\":\"DeleteRequest\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody\"},{\"kind\":1024,\"name\":\"controller\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.controller\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"property\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.property\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"environment\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.environment\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"regulation\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.regulation\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"jurisdiction\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.jurisdiction\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"identities\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.identities\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"callbacks\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.callbacks\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"subject\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.subject\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"context\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.context\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.context.__type-1\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type.context\"},{\"kind\":1024,\"name\":\"submittedTimestamp\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.submittedTimestamp\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type\"},{\"kind\":1024,\"name\":\"dueTimestamp\",\"url\":\"interfaces/DeleteRequest.html#requestBody.__type.dueTimestamp\",\"classes\":\"\",\"parent\":\"DeleteRequest.requestBody.__type\"},{\"kind\":256,\"name\":\"DeleteResponseBody\",\"url\":\"interfaces/DeleteResponseBody.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"status\",\"url\":\"interfaces/DeleteResponseBody.html#status\",\"classes\":\"\",\"parent\":\"DeleteResponseBody\"},{\"kind\":1024,\"name\":\"reason\",\"url\":\"interfaces/DeleteResponseBody.html#reason\",\"classes\":\"\",\"parent\":\"DeleteResponseBody\"},{\"kind\":1024,\"name\":\"expectedCompletionTimestamp\",\"url\":\"interfaces/DeleteResponseBody.html#expectedCompletionTimestamp\",\"classes\":\"\",\"parent\":\"DeleteResponseBody\"},{\"kind\":1024,\"name\":\"redirectUrl\",\"url\":\"interfaces/DeleteResponseBody.html#redirectUrl\",\"classes\":\"\",\"parent\":\"DeleteResponseBody\"},{\"kind\":1024,\"name\":\"requestId\",\"url\":\"interfaces/DeleteResponseBody.html#requestId\",\"classes\":\"\",\"parent\":\"DeleteResponseBody\"},{\"kind\":1024,\"name\":\"documents\",\"url\":\"interfaces/DeleteResponseBody.html#documents\",\"classes\":\"\",\"parent\":\"DeleteResponseBody\"},{\"kind\":1024,\"name\":\"context\",\"url\":\"interfaces/DeleteResponseBody.html#context\",\"classes\":\"\",\"parent\":\"DeleteResponseBody\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/DeleteResponseBody.html#context.__type\",\"classes\":\"\",\"parent\":\"DeleteResponseBody.context\"},{\"kind\":1024,\"name\":\"subject\",\"url\":\"interfaces/DeleteResponseBody.html#subject\",\"classes\":\"\",\"parent\":\"DeleteResponseBody\"},{\"kind\":1024,\"name\":\"identities\",\"url\":\"interfaces/DeleteResponseBody.html#identities\",\"classes\":\"\",\"parent\":\"DeleteResponseBody\"},{\"kind\":256,\"name\":\"DeleteResponse\",\"url\":\"interfaces/DeleteResponse.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/DeleteResponse.html#apiVersion\",\"classes\":\"\",\"parent\":\"DeleteResponse\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/DeleteResponse.html#kind\",\"classes\":\"\",\"parent\":\"DeleteResponse\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/DeleteResponse.html#metadata\",\"classes\":\"\",\"parent\":\"DeleteResponse\"},{\"kind\":1024,\"name\":\"response\",\"url\":\"interfaces/DeleteResponse.html#response\",\"classes\":\"\",\"parent\":\"DeleteResponse\"},{\"kind\":256,\"name\":\"DeleteStatusEvent\",\"url\":\"interfaces/DeleteStatusEvent.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"apiVersion\",\"url\":\"interfaces/DeleteStatusEvent.html#apiVersion\",\"classes\":\"\",\"parent\":\"DeleteStatusEvent\"},{\"kind\":1024,\"name\":\"kind\",\"url\":\"interfaces/DeleteStatusEvent.html#kind\",\"classes\":\"\",\"parent\":\"DeleteStatusEvent\"},{\"kind\":1024,\"name\":\"metadata\",\"url\":\"interfaces/DeleteStatusEvent.html#metadata\",\"classes\":\"\",\"parent\":\"DeleteStatusEvent\"},{\"kind\":1024,\"name\":\"event\",\"url\":\"interfaces/DeleteStatusEvent.html#event\",\"classes\":\"\",\"parent\":\"DeleteStatusEvent\"},{\"kind\":4194304,\"name\":\"DsrRequest\",\"url\":\"types/DsrRequest.html\",\"classes\":\"\"},{\"kind\":4194304,\"name\":\"DsrResponse\",\"url\":\"types/DsrResponse.html\",\"classes\":\"\"},{\"kind\":256,\"name\":\"ConnectionConfig\",\"url\":\"interfaces/ConnectionConfig.html\",\"classes\":\"\"}],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[[\"name/0\",[0,45.574]],[\"comment/0\",[]],[\"name/1\",[1,26.115]],[\"comment/1\",[]],[\"name/2\",[2,45.574]],[\"comment/2\",[]],[\"name/3\",[3,45.574]],[\"comment/3\",[]],[\"name/4\",[4,45.574]],[\"comment/4\",[]],[\"name/5\",[5,45.574]],[\"comment/5\",[]],[\"name/6\",[6,45.574]],[\"comment/6\",[]],[\"name/7\",[7,45.574]],[\"comment/7\",[]],[\"name/8\",[8,45.574]],[\"comment/8\",[]],[\"name/9\",[9,45.574]],[\"comment/9\",[]],[\"name/10\",[10,45.574]],[\"comment/10\",[]],[\"name/11\",[11,45.574]],[\"comment/11\",[]],[\"name/12\",[12,45.574]],[\"comment/12\",[]],[\"name/13\",[13,45.574]],[\"comment/13\",[]],[\"name/14\",[14,45.574]],[\"comment/14\",[]],[\"name/15\",[15,45.574]],[\"comment/15\",[]],[\"name/16\",[16,45.574]],[\"comment/16\",[]],[\"name/17\",[17,45.574]],[\"comment/17\",[]],[\"name/18\",[18,45.574]],[\"comment/18\",[]],[\"name/19\",[19,45.574]],[\"comment/19\",[]],[\"name/20\",[20,45.574]],[\"comment/20\",[]],[\"name/21\",[21,45.574]],[\"comment/21\",[]],[\"name/22\",[22,45.574]],[\"comment/22\",[]],[\"name/23\",[23,29.479]],[\"comment/23\",[]],[\"name/24\",[24,40.466]],[\"comment/24\",[]],[\"name/25\",[25,40.466]],[\"comment/25\",[]],[\"name/26\",[26,40.466]],[\"comment/26\",[]],[\"name/27\",[27,40.466]],[\"comment/27\",[]],[\"name/28\",[28,40.466]],[\"comment/28\",[]],[\"name/29\",[29,40.466]],[\"comment/29\",[]],[\"name/30\",[30,45.574]],[\"comment/30\",[]],[\"name/31\",[31,45.574]],[\"comment/31\",[]],[\"name/32\",[32,45.574]],[\"comment/32\",[]],[\"name/33\",[33,45.574]],[\"comment/33\",[]],[\"name/34\",[34,45.574]],[\"comment/34\",[]],[\"name/35\",[35,40.466]],[\"comment/35\",[]],[\"name/36\",[36,40.466]],[\"comment/36\",[]],[\"name/37\",[1,26.115]],[\"comment/37\",[]],[\"name/38\",[37,45.574]],[\"comment/38\",[]],[\"name/39\",[38,45.574]],[\"comment/39\",[]],[\"name/40\",[35,40.466]],[\"comment/40\",[]],[\"name/41\",[36,40.466]],[\"comment/41\",[]],[\"name/42\",[1,26.115]],[\"comment/42\",[]],[\"name/43\",[39,45.574]],[\"comment/43\",[]],[\"name/44\",[40,45.574]],[\"comment/44\",[]],[\"name/45\",[41,45.574]],[\"comment/45\",[]],[\"name/46\",[42,45.574]],[\"comment/46\",[]],[\"name/47\",[43,45.574]],[\"comment/47\",[]],[\"name/48\",[44,45.574]],[\"comment/48\",[]],[\"name/49\",[45,45.574]],[\"comment/49\",[]],[\"name/50\",[46,45.574]],[\"comment/50\",[]],[\"name/51\",[47,45.574]],[\"comment/51\",[]],[\"name/52\",[48,45.574]],[\"comment/52\",[]],[\"name/53\",[49,45.574]],[\"comment/53\",[]],[\"name/54\",[50,45.574]],[\"comment/54\",[]],[\"name/55\",[51,45.574]],[\"comment/55\",[]],[\"name/56\",[52,45.574]],[\"comment/56\",[]],[\"name/57\",[1,26.115]],[\"comment/57\",[]],[\"name/58\",[53,29.479]],[\"comment/58\",[]],[\"name/59\",[54,45.574]],[\"comment/59\",[]],[\"name/60\",[55,45.574]],[\"comment/60\",[]],[\"name/61\",[24,40.466]],[\"comment/61\",[]],[\"name/62\",[56,30.91]],[\"comment/62\",[]],[\"name/63\",[23,29.479]],[\"comment/63\",[]],[\"name/64\",[53,29.479]],[\"comment/64\",[]],[\"name/65\",[57,40.466]],[\"comment/65\",[]],[\"name/66\",[1,26.115]],[\"comment/66\",[]],[\"name/67\",[58,40.466]],[\"comment/67\",[]],[\"name/68\",[59,40.466]],[\"comment/68\",[]],[\"name/69\",[60,40.466]],[\"comment/69\",[]],[\"name/70\",[61,40.466]],[\"comment/70\",[]],[\"name/71\",[62,40.466]],[\"comment/71\",[]],[\"name/72\",[63,34.588]],[\"comment/72\",[]],[\"name/73\",[64,40.466]],[\"comment/73\",[]],[\"name/74\",[65,34.588]],[\"comment/74\",[]],[\"name/75\",[66,34.588]],[\"comment/75\",[]],[\"name/76\",[1,26.115]],[\"comment/76\",[]],[\"name/77\",[67,40.466]],[\"comment/77\",[]],[\"name/78\",[68,40.466]],[\"comment/78\",[]],[\"name/79\",[69,45.574]],[\"comment/79\",[]],[\"name/80\",[70,40.466]],[\"comment/80\",[]],[\"name/81\",[71,40.466]],[\"comment/81\",[]],[\"name/82\",[72,40.466]],[\"comment/82\",[]],[\"name/83\",[73,40.466]],[\"comment/83\",[]],[\"name/84\",[74,40.466]],[\"comment/84\",[]],[\"name/85\",[75,40.466]],[\"comment/85\",[]],[\"name/86\",[66,34.588]],[\"comment/86\",[]],[\"name/87\",[1,26.115]],[\"comment/87\",[]],[\"name/88\",[65,34.588]],[\"comment/88\",[]],[\"name/89\",[63,34.588]],[\"comment/89\",[]],[\"name/90\",[25,40.466]],[\"comment/90\",[]],[\"name/91\",[56,30.91]],[\"comment/91\",[]],[\"name/92\",[23,29.479]],[\"comment/92\",[]],[\"name/93\",[53,29.479]],[\"comment/93\",[]],[\"name/94\",[76,45.574]],[\"comment/94\",[]],[\"name/95\",[26,40.466]],[\"comment/95\",[]],[\"name/96\",[56,30.91]],[\"comment/96\",[]],[\"name/97\",[23,29.479]],[\"comment/97\",[]],[\"name/98\",[53,29.479]],[\"comment/98\",[]],[\"name/99\",[77,40.466]],[\"comment/99\",[]],[\"name/100\",[27,40.466]],[\"comment/100\",[]],[\"name/101\",[56,30.91]],[\"comment/101\",[]],[\"name/102\",[23,29.479]],[\"comment/102\",[]],[\"name/103\",[53,29.479]],[\"comment/103\",[]],[\"name/104\",[57,40.466]],[\"comment/104\",[]],[\"name/105\",[1,26.115]],[\"comment/105\",[]],[\"name/106\",[58,40.466]],[\"comment/106\",[]],[\"name/107\",[59,40.466]],[\"comment/107\",[]],[\"name/108\",[60,40.466]],[\"comment/108\",[]],[\"name/109\",[61,40.466]],[\"comment/109\",[]],[\"name/110\",[62,40.466]],[\"comment/110\",[]],[\"name/111\",[63,34.588]],[\"comment/111\",[]],[\"name/112\",[64,40.466]],[\"comment/112\",[]],[\"name/113\",[65,34.588]],[\"comment/113\",[]],[\"name/114\",[66,34.588]],[\"comment/114\",[]],[\"name/115\",[1,26.115]],[\"comment/115\",[]],[\"name/116\",[67,40.466]],[\"comment/116\",[]],[\"name/117\",[68,40.466]],[\"comment/117\",[]],[\"name/118\",[78,45.574]],[\"comment/118\",[]],[\"name/119\",[70,40.466]],[\"comment/119\",[]],[\"name/120\",[71,40.466]],[\"comment/120\",[]],[\"name/121\",[72,40.466]],[\"comment/121\",[]],[\"name/122\",[73,40.466]],[\"comment/122\",[]],[\"name/123\",[74,40.466]],[\"comment/123\",[]],[\"name/124\",[75,40.466]],[\"comment/124\",[]],[\"name/125\",[66,34.588]],[\"comment/125\",[]],[\"name/126\",[1,26.115]],[\"comment/126\",[]],[\"name/127\",[65,34.588]],[\"comment/127\",[]],[\"name/128\",[63,34.588]],[\"comment/128\",[]],[\"name/129\",[28,40.466]],[\"comment/129\",[]],[\"name/130\",[56,30.91]],[\"comment/130\",[]],[\"name/131\",[23,29.479]],[\"comment/131\",[]],[\"name/132\",[53,29.479]],[\"comment/132\",[]],[\"name/133\",[79,45.574]],[\"comment/133\",[]],[\"name/134\",[29,40.466]],[\"comment/134\",[]],[\"name/135\",[56,30.91]],[\"comment/135\",[]],[\"name/136\",[23,29.479]],[\"comment/136\",[]],[\"name/137\",[53,29.479]],[\"comment/137\",[]],[\"name/138\",[77,40.466]],[\"comment/138\",[]],[\"name/139\",[80,45.574]],[\"comment/139\",[]],[\"name/140\",[81,45.574]],[\"comment/140\",[]],[\"name/141\",[82,45.574]],[\"comment/141\",[]]],\"invertedIndex\":[[\"__type\",{\"_index\":1,\"name\":{\"1\":{},\"37\":{},\"42\":{},\"57\":{},\"66\":{},\"76\":{},\"87\":{},\"105\":{},\"115\":{},\"126\":{}},\"comment\":{}}],[\"accessrequest\",{\"_index\":24,\"name\":{\"24\":{},\"61\":{}},\"comment\":{}}],[\"accessresponse\",{\"_index\":25,\"name\":{\"25\":{},\"90\":{}},\"comment\":{}}],[\"accessresponsebody\",{\"_index\":69,\"name\":{\"79\":{}},\"comment\":{}}],[\"accessstatusevent\",{\"_index\":26,\"name\":{\"26\":{},\"95\":{}},\"comment\":{}}],[\"addressline1\",{\"_index\":44,\"name\":{\"48\":{}},\"comment\":{}}],[\"addressline2\",{\"_index\":45,\"name\":{\"49\":{}},\"comment\":{}}],[\"apiversion\",{\"_index\":56,\"name\":{\"62\":{},\"91\":{},\"96\":{},\"101\":{},\"130\":{},\"135\":{}},\"comment\":{}}],[\"callback\",{\"_index\":34,\"name\":{\"34\":{}},\"comment\":{}}],[\"callbacks\",{\"_index\":64,\"name\":{\"73\":{},\"112\":{}},\"comment\":{}}],[\"cancelledrequeststatus\",{\"_index\":8,\"name\":{\"8\":{}},\"comment\":{}}],[\"city\",{\"_index\":46,\"name\":{\"50\":{}},\"comment\":{}}],[\"claimnotcoveredrequeststatusreason\",{\"_index\":19,\"name\":{\"19\":{}},\"comment\":{}}],[\"completedrequeststatus\",{\"_index\":7,\"name\":{\"7\":{}},\"comment\":{}}],[\"connectionconfig\",{\"_index\":82,\"name\":{\"141\":{}},\"comment\":{}}],[\"context\",{\"_index\":66,\"name\":{\"75\":{},\"86\":{},\"114\":{},\"125\":{}},\"comment\":{}}],[\"controller\",{\"_index\":58,\"name\":{\"67\":{},\"106\":{}},\"comment\":{}}],[\"countrycode\",{\"_index\":49,\"name\":{\"53\":{}},\"comment\":{}}],[\"data\",{\"_index\":38,\"name\":{\"39\":{}},\"comment\":{}}],[\"datasubject\",{\"_index\":39,\"name\":{\"43\":{}},\"comment\":{}}],[\"deleterequest\",{\"_index\":27,\"name\":{\"27\":{},\"100\":{}},\"comment\":{}}],[\"deleteresponse\",{\"_index\":28,\"name\":{\"28\":{},\"129\":{}},\"comment\":{}}],[\"deleteresponsebody\",{\"_index\":78,\"name\":{\"118\":{}},\"comment\":{}}],[\"deletestatusevent\",{\"_index\":29,\"name\":{\"29\":{},\"134\":{}},\"comment\":{}}],[\"deniedrequeststatus\",{\"_index\":9,\"name\":{\"9\":{}},\"comment\":{}}],[\"description\",{\"_index\":50,\"name\":{\"54\":{}},\"comment\":{}}],[\"document\",{\"_index\":37,\"name\":{\"38\":{}},\"comment\":{}}],[\"documents\",{\"_index\":75,\"name\":{\"85\":{},\"124\":{}},\"comment\":{}}],[\"dsrrequest\",{\"_index\":80,\"name\":{\"139\":{}},\"comment\":{}}],[\"dsrresponse\",{\"_index\":81,\"name\":{\"140\":{}},\"comment\":{}}],[\"dsrversion\",{\"_index\":2,\"name\":{\"2\":{}},\"comment\":{}}],[\"duetimestamp\",{\"_index\":68,\"name\":{\"78\":{},\"117\":{}},\"comment\":{}}],[\"email\",{\"_index\":41,\"name\":{\"45\":{}},\"comment\":{}}],[\"environment\",{\"_index\":60,\"name\":{\"69\":{},\"108\":{}},\"comment\":{}}],[\"event\",{\"_index\":77,\"name\":{\"99\":{},\"138\":{}},\"comment\":{}}],[\"executed\",{\"_index\":15,\"name\":{\"15\":{}},\"comment\":{}}],[\"expectedcompletiontimestamp\",{\"_index\":72,\"name\":{\"82\":{},\"121\":{}},\"comment\":{}}],[\"firstname\",{\"_index\":42,\"name\":{\"46\":{}},\"comment\":{}}],[\"formdata\",{\"_index\":52,\"name\":{\"56\":{}},\"comment\":{}}],[\"handlerequest\",{\"_index\":0,\"name\":{\"0\":{}},\"comment\":{}}],[\"headers\",{\"_index\":36,\"name\":{\"36\":{},\"41\":{}},\"comment\":{}}],[\"identities\",{\"_index\":63,\"name\":{\"72\":{},\"89\":{},\"111\":{},\"128\":{}},\"comment\":{}}],[\"identity\",{\"_index\":30,\"name\":{\"30\":{}},\"comment\":{}}],[\"identityformat\",{\"_index\":32,\"name\":{\"32\":{}},\"comment\":{}}],[\"identityspace\",{\"_index\":31,\"name\":{\"31\":{}},\"comment\":{}}],[\"identityvalue\",{\"_index\":33,\"name\":{\"33\":{}},\"comment\":{}}],[\"inprogressrequeststatus\",{\"_index\":6,\"name\":{\"6\":{}},\"comment\":{}}],[\"insufficientidentification\",{\"_index\":14,\"name\":{\"14\":{}},\"comment\":{}}],[\"insufficientverificationrequeststatusreason\",{\"_index\":17,\"name\":{\"17\":{}},\"comment\":{}}],[\"jurisdiction\",{\"_index\":62,\"name\":{\"71\":{},\"110\":{}},\"comment\":{}}],[\"kind\",{\"_index\":23,\"name\":{\"23\":{},\"63\":{},\"92\":{},\"97\":{},\"102\":{},\"131\":{},\"136\":{}},\"comment\":{}}],[\"lastname\",{\"_index\":43,\"name\":{\"47\":{}},\"comment\":{}}],[\"metadata\",{\"_index\":53,\"name\":{\"58\":{},\"64\":{},\"93\":{},\"98\":{},\"103\":{},\"132\":{},\"137\":{}},\"comment\":{}}],[\"needuserverificationrequeststatusreason\",{\"_index\":12,\"name\":{\"12\":{}},\"comment\":{}}],[\"nomatchrequeststatusreason\",{\"_index\":18,\"name\":{\"18\":{}},\"comment\":{}}],[\"otherrequeststatusreason\",{\"_index\":22,\"name\":{\"22\":{}},\"comment\":{}}],[\"outsidejurisdictionrequeststatusreason\",{\"_index\":20,\"name\":{\"20\":{}},\"comment\":{}}],[\"pendingrequeststatus\",{\"_index\":5,\"name\":{\"5\":{}},\"comment\":{}}],[\"phone\",{\"_index\":51,\"name\":{\"55\":{}},\"comment\":{}}],[\"postalcode\",{\"_index\":48,\"name\":{\"52\":{}},\"comment\":{}}],[\"property\",{\"_index\":59,\"name\":{\"68\":{},\"107\":{}},\"comment\":{}}],[\"reason\",{\"_index\":71,\"name\":{\"81\":{},\"120\":{}},\"comment\":{}}],[\"redirecturl\",{\"_index\":73,\"name\":{\"83\":{},\"122\":{}},\"comment\":{}}],[\"regulation\",{\"_index\":61,\"name\":{\"70\":{},\"109\":{}},\"comment\":{}}],[\"requestbody\",{\"_index\":57,\"name\":{\"65\":{},\"104\":{}},\"comment\":{}}],[\"requested\",{\"_index\":13,\"name\":{\"13\":{}},\"comment\":{}}],[\"requestid\",{\"_index\":74,\"name\":{\"84\":{},\"123\":{}},\"comment\":{}}],[\"requeststatus\",{\"_index\":3,\"name\":{\"3\":{}},\"comment\":{}}],[\"requeststatusreason\",{\"_index\":10,\"name\":{\"10\":{}},\"comment\":{}}],[\"response\",{\"_index\":79,\"name\":{\"133\":{}},\"comment\":{}}],[\"responsebody\",{\"_index\":76,\"name\":{\"94\":{}},\"comment\":{}}],[\"stateregioncode\",{\"_index\":47,\"name\":{\"51\":{}},\"comment\":{}}],[\"status\",{\"_index\":70,\"name\":{\"80\":{},\"119\":{}},\"comment\":{}}],[\"subject\",{\"_index\":65,\"name\":{\"74\":{},\"88\":{},\"113\":{},\"127\":{}},\"comment\":{}}],[\"submittedtimestamp\",{\"_index\":67,\"name\":{\"77\":{},\"116\":{}},\"comment\":{}}],[\"suspectedfraudrequeststatusreason\",{\"_index\":16,\"name\":{\"16\":{}},\"comment\":{}}],[\"tenant\",{\"_index\":55,\"name\":{\"60\":{}},\"comment\":{}}],[\"toomanyrequestsrequeststatusreason\",{\"_index\":21,\"name\":{\"21\":{}},\"comment\":{}}],[\"type\",{\"_index\":40,\"name\":{\"44\":{}},\"comment\":{}}],[\"uid\",{\"_index\":54,\"name\":{\"59\":{}},\"comment\":{}}],[\"unknownrequeststatus\",{\"_index\":4,\"name\":{\"4\":{}},\"comment\":{}}],[\"unknownrequeststatusreason\",{\"_index\":11,\"name\":{\"11\":{}},\"comment\":{}}],[\"url\",{\"_index\":35,\"name\":{\"35\":{},\"40\":{}},\"comment\":{}}]],\"pipeline\":[]}}");
